refactor: separate controlled config from dynamic config
